requirements.txt
crewai>=0.16.11
python-dotenv
pyyaml
